+++ test/shell.escape.js
@@ -1,6 +1,6 @@
-import shell       from '..'
-import test        from 'ava'
-import { inspect } from 'util'
+const test        = require('ava')
+const { inspect } = require('util')
+const shell       = require('..')
 
 test('stringifies to the escaped string', t => {
     t.is(shell.escape('Foo Bar').toString(), "'Foo Bar'")
@@ -13,7 +13,7 @@ test('escapes a string with spaces', t => {
 
 test('escapes an array of strings with spaces', t => {
     t.is(
-        shell.escape([ 'Foo Bar', 'Baz Quux' ]).value,
+        shell.escape(['Foo Bar', 'Baz Quux']).value,
         "'Foo Bar' 'Baz Quux'"
     )
 })
@@ -27,7 +27,7 @@ test('escapes a string with quotes', t => {
 
 test('escapes an array of strings with quotes', t => {
     t.is(
-        shell.escape([ `Foo's "Bar"`, `Foo 'Bar' "Baz"` ]).value,
+        shell.escape([`Foo's "Bar"`, `Foo 'Bar' "Baz"`]).value,
         `'Foo'"'"'s "Bar"' 'Foo '"'"'Bar'"'"' "Baz"'`
     )
 })
@@ -50,7 +50,7 @@ test('ignores null and undefined array values', t => {
 })
 
 test('stringifies defined values', t => {
-    t.is(shell.escape('').value, '')
+    t.is(shell.escape('').value, "''")
     t.is(shell.escape(false).value, 'false')
     t.is(shell.escape(42).value, '42')
 })
@@ -64,7 +64,7 @@ test('stringifies defined array values', t => {
             false,
             'Baz Quux'
         ]).value,
-        `'Foo Bar' 0  false 'Baz Quux'`
+        `'Foo Bar' 0 '' false 'Baz Quux'`
     )
 })
 
@@ -74,12 +74,12 @@ test('flattens nested array values', t => {
             [ 'Foo Bar',
                 [ 0, '', false,
                     [ null, undefined,
-                        [ 'Baz Quux' ]
+                        ['Baz Quux']
                     ]
                 ]
             ]
         ]).value,
-        `'Foo Bar' 0  false 'Baz Quux'`
+        `'Foo Bar' 0 '' false 'Baz Quux'`
     )
 })
 
@@ -89,7 +89,7 @@ test("doesn't escape embedded escaped strings", t => {
 })
 
 test("doesn't escape embedded escaped arrays", t => {
-    const escaped = shell.escape([ 'Foo Bar', 'Baz Quux' ])
+    const escaped = shell.escape(['Foo Bar', 'Baz Quux'])
     t.is(shell`foo ${escaped}`, "foo 'Foo Bar' 'Baz Quux'")
 })
 
@@ -97,7 +97,7 @@ test("doesn't escape embedded nested strings/arrays", t => {
     const escaped = shell.escape([
         '1 2',
         shell.escape('3 4'),
-        shell.escape([ '5 6', '7 8' ]),
+        shell.escape(['5 6', '7 8']),
         '9 10'
     ])
 
@@ -108,7 +108,7 @@ test("doesn't escape embedded nested strings/arrays", t => {
 })
 
 test('supports embedded preserves', t => {
-    const param = shell.escape(shell.preserve('foo bar'), [ "baz's quux" ])
+    const param = shell.escape(shell.preserve('foo bar'), ["baz's quux"])
 
     t.is(
         shell`command ${param}`,